refactor helper out into separate function
